Extract 2mp scraping
