+++ slash_commands/index-2mp.js
@@ -10,24 +10,13 @@ const { MessageActionRow, MessageButton } = require('discord.js');
 
 const Parsed = require('../parser/parsed.js');
 
-const COLS = {
-    NUMBER: 'B',
-    ENTITY: 'C',
-    UPGRADE: 'E',
-    OG_MAP: 'F',
-    VERSION: 'H',
-    DATE: 'I',
-    PERSON: 'J',
-    LINK: 'L',
-};
+const { COLS } = require('../services/index/2mp_scraper')
 
 const TOWER_COLS = {
     TOWER: 'O',
     BASE: 'P',
     LAST: 'Y',
 };
-
-CACHE_FNAME_2MP = '2mp.json'
 
 const { 
     SlashCommandBuilder, 
@@ -93,15 +82,9 @@ async function execute(interaction) {
 
     const forceReload = interaction.options.getString('reload') ? true : false
 
-    let allCombos;
-    if (Index.hasCachedCombos(CACHE_FNAME_2MP) && !forceReload) {
-        allCombos = await Index.fetchCachedCombos(CACHE_FNAME_2MP)           
-    } else {
-        allCombos = await scrapeAllCombos();
-        Index.cacheCombos(allCombos, CACHE_FNAME_2MP)
-    }
+    const allCombos = await Index.fetchCombos('2mp', reload=forceReload)
 
-    const mtime = Index.getLastCacheModified(CACHE_FNAME_2MP)
+    const mtime = Index.getLastCacheModified('2mp')
 
     if ((parsed.tower_upgrade || parsed.hero) && !parsed.person) {
         const entity = parsed.hero || Towers.towerUpgradeToIndexNormalForm(parsed.tower_upgrade)
@@ -219,101 +202,6 @@ function parsePerson(interaction) {
     if (person) {
         return CommandParser.parse([`user#${person}`], new PersonParser())
     } else return new Parsed();
-}
-
-////////////////////////////////////////////////////////////
-// Parsing Combos (when reloading)
-////////////////////////////////////////////////////////////
-
-async function scrapeAllCombos() {
-    const sheet = GoogleSheetsHelper.sheetByName(Btd6Index, '2mpc');
-
-    // Load ENTITY and MAP columns
-    let [startRow, endRow] = await rowBoundaries();
-    await sheet.loadCells(`${COLS.NUMBER}${startRow}:${COLS.LINK}${endRow}`);
-
-    let combos = []
-    // Retrieve og- and alt-map notes from each tower row
-    for (var row = startRow; row <= endRow; row++) {
-        combos.push(
-            parsePreloadedRow(row)
-        );
-    }
-
-    return combos;
-}
-
-async function rowBoundaries() {
-    const sheet = GoogleSheetsHelper.sheetByName(Btd6Index, '2mpc');
-    await sheet.loadCells(`${COLS.NUMBER}1:${COLS.NUMBER}${sheet.rowCount}`);
-
-    startRow = null;
-    endRow = null;
-
-    for (let row = 1; row <= sheet.rowCount; row++) {
-        numberCandidate = sheet.getCellByA1(`${COLS.NUMBER}${row}`).value;
-
-        // If startRow has been found, find the first occurence of a blank cell
-        if (startRow && !numberCandidate) {
-            endRow = row - 1;
-            break;
-        } else if (
-            numberCandidate &&
-            numberCandidate.replace(/ /g, '') === '1st'
-        ) {
-            startRow = row;
-        }
-    }
-
-    if (!startRow) {
-        throw new DeveloperCommandError(
-            `Orientation failed because \`1st\` couldn't be found in the "Number" column.`
-        );
-    }
-
-    // If there wasn't a trailing blank cell, then the last cell viewed must be the end cell
-    if (!endRow) endRow = sheet.rowCount;
-
-    return [startRow, endRow];
-}
-
-// Assumes appropriate cells are loaded beforehand!
-function parsePreloadedRow(row) {
-    const sheet = GoogleSheetsHelper.sheetByName(Btd6Index, '2mpc');
-
-    let completion = {}
-    for (const col of ["NUMBER", "ENTITY", "UPGRADE", "VERSION"]) {
-        completion[col] = sheet.getCellByA1(`${COLS[col]}${row}`).value;
-    }
-    completion.DATE = sheet.getCellByA1(`${COLS.DATE}${row}`).formattedValue;
-
-    completion.MAPS = parseMapCompletions(row);
-
-    return completion
-}
-
-function parseMapCompletions(row) {
-    const sheet = GoogleSheetsHelper.sheetByName(Btd6Index, '2mpc');
-
-    const ogMapCell = sheet.getCellByA1(`${COLS.OG_MAP}${row}`);
-    const ogMapAbbr = Aliases.mapToIndexAbbreviation(
-        Aliases.toAliasNormalForm(ogMapCell.value)
-    );
-    const ogPerson = sheet.getCellByA1(`${COLS.PERSON}${row}`).value;
-    const ogLinkCell = sheet.getCellByA1(`${COLS.LINK}${row}`);
-
-    const ogMapCompletion = {
-        PERSON: ogPerson,
-        LINK: `[${ogLinkCell.value}](${ogLinkCell.hyperlink})`,
-        OG: true,
-    };
-    // Add rest of maps found in notes
-    const maps = {
-        [ogMapAbbr]: ogMapCompletion,
-        ...Index.parseMapNotes(ogMapCell.note),
-    };
-
-    return maps
 }
 
 ////////////////////////////////////////////////////////////
@@ -764,6 +652,8 @@ async function embedOneOrMultiplePages(interaction, title, columns, numOGComplet
 ////////////////////////////////////////////////////////////
 // Tower Statistics
 ////////////////////////////////////////////////////////////
+
+// TODO: Extract parsing to 2mp_scraper
 
 // Displays a 3x3 grid completion checkboxes/x'es for each upgrade+tier
 // Displays base centered above grid
